Refactor application structure and update roadmap status
- Reorganized project structure: specific folders for Views (Editors, Layout, Explorer, Settings).
- Implemented core Services: FileOperationsService, WorkspaceService.
- Added SpreadsheetGrid control and related ViewModels.
- Completed Phases 1-6 in ROADMAP.md; Phase 7 is In Progress.
- Refined UI components and Agent interactions.

+++ core/Agents/JsonAgent.cs
@@ -218,9 +218,9 @@ public class JsonAgent
     }
 }
 
-public record JsonParseResult(bool IsValid, JsonModel? Model, string? ErrorMessage, int? LineNumber)
+public record JsonParseResult(bool IsValid, JsonModel? Model, string? ErrorMessage, long? LineNumber)
 {
     public static JsonParseResult Success(JsonModel model) => new(true, model, null, null);
 
-    public static JsonParseResult Failure(string message, int? lineNumber = null) => new(false, null, message, lineNumber);
+    public static JsonParseResult Failure(string message, long? lineNumber = null) => new(false, null, message, lineNumber);
 }
